Remove -x from xargs invocation again
I *thought* that this would cause xargs to exit if the command line gets
too long. This is not the case. The reason for xargs existence is
precisely to avoid the problem of things crashing if argument lists get
too long.

If we really wanted to crash in this situation, we should do the naive
way of passing argument lists which is using borg create ... $(cat
sourcedirs)